Add layout for demo'ing recovering a fragment [PARTIAL]
This will be useful for when we want to check if a dialog is visible.
In situations like config changes we don't have a handle to the
fragment and we'll have to get one through the fragment manager and
the fragment's tag.
Add initial support for showing the dialog. Eventually it will be
closed after a certain amount of time.